requirements.txt
streamlit==1.28.0
python-binance==1.0.19
pandas==2.1.0
numpy==1.24.0
plotly==5.17.0
scipy==1.11.0
statsmodels==0.14.0
scikit-learn==1.3.0